requirements.txt
dj-static==0.0.6
Django==3.0.5
Pillow==7.0.0
django-tastypie==0.14.3

geojson==2.5.0
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2019.3
six==1.14.0
static3==0.7.0
psycopg2-binary==2.8.4
djangorestframework==3.11.0
markdown==3.1.1
pypng==0.0.20
requests==2.22.0
pymongo==3.10.1
django-filter==2.2.0
otpauth==1.0.1
qrcode==6.1
cryptography==3.0
openpyxl==3.0.4